rustchain-bounty-agent/requirements.txt
anthropic>=0.18.0
PyGithub>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
